+++ src/Pages/HomePage/HomePageStyle.js
@@ -34,6 +34,9 @@ export default StyleSheet.create({
   },
   postOption: {
     paddingVertical: 12,
+    flexDirection: "row",
+    alignItems: "center",
+    gap: 10,
   },
   postOptionText: {
     fontSize: 16,